Corrected examples for new io.Writer specification
